cdk/requirements.txt
aws-cdk.core==1.16.3
aws-cdk.aws_ec2==1.16.3
aws-cdk.aws_ecs==1.16.3
aws-cdk.aws_servicediscovery==1.16.3
